requirements.txt
discord-py
asyncio
datetime
Cybernator